updating LCD in main loop; MDNS workaround; ADC enhancement

+++ src/MqttRemoteControl.cpp
@@ -383,7 +383,7 @@ void MqttRemoteControl::_onModeChange(char* payload,size_t len){
  
     #if SerialDebug
     DBG_PRINTF("MQTT:mode path value:");
-    for(int i=0;i<len;i++)
+    for(size_t i=0;i<len;i++)
         DBG_PRINTF("%c",payload[i]);
     DBG_PRINTF("\n");
     #endif
